fix: ocr processing
