+++ services/ocr/OCR/errors.py
@@ -1,0 +1,6 @@
+
+class QRCodeError(Exception):
+    pass
+
+class NamesDetectionError(Exception):
+    pass
