+++ services/ocr/main.py
@@ -1,10 +1,10 @@
 
-from ocr import Engine
+from OCR.engine import Engine
 import numpy as np
 import pika, sys, os, json, base64, cv2, argparse
 
-parser = argparse.ArgumentParser(description='OCR service test')
-parser.add_argument('--docker_mode', default=False, type=bool, help='Is in docker?')
+parser = argparse.ArgumentParser(description='OCR service')
+parser.add_argument('--docker_mode', action="store_true")
 args = parser.parse_args()
 
 if args.docker_mode: RABBITMQ_HOST = "rabbitmq"
@@ -12,6 +12,7 @@ else: RABBITMQ_HOST = "127.0.0.1"
 
 def main():
     connection = pika.BlockingConnection(pika.ConnectionParameters(host=RABBITMQ_HOST))
+    connection2 = pika.BlockingConnection(pika.ConnectionParameters(host=RABBITMQ_HOST))
     channel = connection.channel()
 
     def callback(ch, method, properties, body):
@@ -19,12 +20,12 @@ def main():
         nparr = np.frombuffer(base64.b64decode(data["img"]), np.uint8)
         img = cv2.imdecode(nparr, cv2.IMREAD_COLOR)
 
-        out = Engine.process(img, int(data["week_number"]), RABBITMQ_HOST, data["id"])
+        out = Engine.process(img, int(data["week_number"]), connection2, data["id"])
 
         channel.queue_declare(queue=f"scan:shift")
         channel.basic_publish(exchange='',
                             routing_key=f"scan:shift",
-                            body=json.dumps({"status": "DONE", "scan_id": data["id"]}))
+                            body=json.dumps({"status": "PROCCESED", "scan_id": data["id"]}))
 
     channel.basic_consume(queue='ocr-queue', on_message_callback=callback, auto_ack=True)
